+++ xmchord/model/action_reader.cc
@@ -205,15 +205,23 @@ std::string ActionReader::WrapOutputLine(const std::string &line) {
   auto segments = helper::Textual::Explode(line, ' ');
 
   std::string last_line_wrapped;
+  size_t len_chording_identifier = 0;
   uint16_t current_line_length = 0;
 
   for (const auto& segment : segments) {
     auto segment_len = segment.length();
 
     if (current_line_length + segment_len > 80) {
+      if (len_chording_identifier == 0) {
+        len_chording_identifier = line.find(" -");
+
+        if (len_chording_identifier == std::string::npos)
+          len_chording_identifier = len_longest_filename_ + 3;
+      }
+
       last_line_wrapped +=
           "\n"
-          + helper::Textual::Repeat(" ", len_longest_filename_ + 6)
+          + helper::Textual::Repeat(" ", len_chording_identifier)
           + segment;
 
       current_line_length = segment_len;