+++ bin/actions/0220022-108.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 #: ◢ + Down -
-#: Jump-scroll to bottom- the required key combo differs among programs,
+#: Jump-scroll to bottom- the invoked required key combo differs among programs,
 #: some need "End" others "Ctrl+End" (e.g. LibreOffice)
 
 focusApplication=\