+++ bin/actions/1010101-105.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
-#: ◣ + left -
-#: Within Gnome Terminal: Go Tab to left.
+#: ◣ + Left - Within Gnome Terminal: Go Tab to left.
 #: Within Chromium or Firefox: Go URL back.
 
 focusApplication=\